requirements.txt
fastapi==0.114.0
uvicorn==0.30.6
supabase
pydantic_settings
starlette